requirements.txt
discord
langchain
langchain-community
openai
faiss-cpu
python-dotenv
